improved UI & removed unnecessary code

+++ src/Deed/Deed.Application/Capitals/Commands/Delete/DeleteCapitalCommandHandler.cs
@@ -13,11 +13,16 @@ internal sealed class DeleteCapitalCommandHandler(
 {
     public async Task<Result> Handle(DeleteCapitalCommand command, CancellationToken cancellationToken)
     {
-        var capital = await repository.GetAsync(new CapitalByIdSpecification(command.Id)).ConfigureAwait(false);
+        var capital = await repository.GetAsync(new CapitalByIdSpecification(command.Id, true, true, true, true)).ConfigureAwait(false);
 
         if (capital is null)
         {
             return Result.Failure(DomainErrors.General.NotFound(nameof(capital)));
+        }
+
+        if (capital.HasReferences())
+        {
+            return Result.Failure(DomainErrors.Capital.ReferenceExists);
         }
 
         repository.Delete(capital);
